src/recall/requirements.txt
json
faiss-cpu
Django
numpy
sentence_transformers

